refactor >> hoverReveal
