+++ src/pages/detailPage/components/similarMovie.js
@@ -1,19 +1,30 @@
 import { useQuery } from "react-query";
 import { getSimilarMovie } from "../../../api";
 import styled from "styled-components";
+import HoverRevealComponents from "../../../components/hoverReveal";
+import { useNavigate, useParams } from "react-router-dom";
 
 const SimilarMovie = ({ movieId }) => {
+  const prams = useParams();
+  const navigate = useNavigate();
+
   const { data } = useQuery(["similar"], () => getSimilarMovie(movieId));
 
   const similarArr = data && data.results;
 
+  // 상세 페이지 이동 함수
+  const onOpenDetailPage = (movie) => {
+    navigate(`/${prams.movie}/:${movie.id}`);
+  };
+
   return (
     data && (
       <SimilarMovieWrapper>
-        {similarArr.map((img, index) => (
-          <MovieImg
+        {similarArr.map((movie, index) => (
+          <HoverRevealComponents
+            movie={movie}
             key={index}
-            src={`https://image.tmdb.org/t/p/w500${img.poster_path}`}
+            onOpenDetailPage={onOpenDetailPage}
           />
         ))}
       </SimilarMovieWrapper>
@@ -27,9 +38,4 @@ const SimilarMovieWrapper = styled.div`
   display: grid;
   grid-template-columns: repeat(5, 1fr);
   grid-row-gap: 50px;
-`;
-
-const MovieImg = styled.img`
-  width: 212px;
-  height: 282px;
 `;
