+++ src/pages/searchResultPage/index.js
@@ -1,12 +1,19 @@
 import { styled } from "styled-components";
 import { flexAlignCenter, flexCenter } from "../../styles/common.style";
 import { useNavigate } from "react-router-dom";
+import { useQuery } from "react-query";
+import { getSearchMovie } from "../../api";
 
 const SearchResultPage = () => {
   const navigate = useNavigate();
   const onClickDetailPage = () => {
     navigate("/detailPage");
   };
+
+  const { data } = useQuery(["qqqqq"], () => getSearchMovie("Avengers"));
+
+  data && console.log(data);
+
   return (
     <>
       <SearchPageWapper>
